Assistant checkpoint: Convert float to int in price calculations
Assistant generated file changes:
- handlers/admin_handlers/xizmat_qoshish.py: Change float to int in xizmat_qoshish_narxi function
- handlers/user_handlers/buyurtma_davomi.py: Change price calculation to use integer division

---

User prompt:

Yana bor, qayerlarda float ishlatilgan

+++ handlers/admin_handlers/xizmat_qoshish.py
@@ -61,7 +61,7 @@ async def xizmat_qoshish_narxi(message: Message, state: FSMContext):
     if int(narxi) < 0:
         await message.answer("Narxi manfiy bo'lishi mumkin emas")
         return
-    narxi = float(narxi)
+    narxi = int(narxi)
     await state.update_data(narxi=narxi)
     await message.answer("Tavsifini kiriting")
     await state.set_state(Xizmat_qosh.tavsif)
